Chatgpt-chatbot_Nextjs_Openaisdk-fastapi/fontend-plus-backend/bk-python/pyproject.toml
[project]
name = "bk-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1",
    "openai-agents>=0.5.0",
    "openai-sdk>=0.1.0",
]
